# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Turismo Ecuador API",
        "description": "API para el manejo de información turística de Ecuador",
        "contact": {
            "email": "soporte@turismo-ec.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/provincias": {
            "get": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Get all provinces",
                "operationId": "09f2d238a699516225ec68b8f9596e30",
                "responses": {
                    "200": {
                        "description": "List of provinces",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Provincia"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Create a new province",
                "operationId": "cd75aa89e80e6b4297db26ad91b57f2a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "codigo_iso",
                                    "region"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Pichincha"
                                    },
                                    "codigo_iso": {
                                        "type": "string",
                                        "example": "P"
                                    },
                                    "region": {
                                        "type": "string",
                                        "example": "Sierra"
                                    },
                                    "poblacion": {
                                        "type": "integer",
                                        "example": 3000000
                                    },
                                    "superficie": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 9416
                                    },
                                    "capital": {
                                        "type": "string",
                                        "example": "Quito"
                                    },
                                    "idioma_principal": {
                                        "type": "string",
                                        "example": "Español"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Province created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provincia"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provincias/{id}": {
            "get": {
                "tags": [
                    "Provincias"
                ],
                "summary": "Get a specific province",
                "operationId": "76806e555a366b1d85afe1e495b4a4b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Province ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Province details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provincia"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Province not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Update a province",
                "operationId": "3c089d9512bc62c513d4c22369c5422d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Province ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Pichincha Actualizada"
                                    },
                                    "codigo_iso": {
                                        "type": "string",
                                        "example": "P"
                                    },
                                    "region": {
                                        "type": "string",
                                        "example": "Sierra"
                                    },
                                    "poblacion": {
                                        "type": "integer",
                                        "example": 3200000
                                    },
                                    "superficie": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 9416
                                    },
                                    "capital": {
                                        "type": "string",
                                        "example": "Quito"
                                    },
                                    "idioma_principal": {
                                        "type": "string",
                                        "example": "Español, Kichwa"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Province updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provincia"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Province not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Delete a province",
                "operationId": "d64805ed478d2c5e6ad7802fc19ec308",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Province ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Province deleted successfully"
                    },
                    "404": {
                        "description": "Province not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Provincia": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Pichincha"
                    },
                    "codigo_iso": {
                        "type": "string",
                        "example": "P"
                    },
                    "region": {
                        "type": "string",
                        "example": "Sierra"
                    },
                    "poblacion": {
                        "type": "integer",
                        "example": 3000000
                    },
                    "superficie": {
                        "type": "number",
                        "format": "float",
                        "example": 9416
                    },
                    "capital": {
                        "type": "string",
                        "example": "Quito"
                    },
                    "idioma_principal": {
                        "type": "string",
                        "example": "Español"
                    },
                    "estado": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}